Perl_sv_backoff - only copy the buffer contents if SvOK(sv)
When undoing an OOK, `Perl_sv_backoff` will always copy the shifted buffer
contents down to the start of the buffer. That's required when the contents
are live, unfortunate when the contents are defunct but tiny, and can have a
noticeably bad effect on performance when the contents are defunct but large.

See https://github.com/Perl/perl5/issues/23967 for an example of the latter.

With this commit the copy will only occur if `SvOK(sv)`. This test is expected
to indicate whether the buffer contents are active or defunct.