Macros guarded by some #ifdef's aren't globally visible
The previous commit undefines macros that aren't supposed to be visible
to XS code.  But, to avoid any possible breakage, it creates an
exception list of symbols that may have been visible, and leaves them
so.  The goal is to stop the list from growing as new code is developed,
and to shorten the list by various means.

This is the first commit to do that, by looking to see if any symbols
aren't actually externally visible because they are guarded by #ifdef's
that evaluate to false.  For example a symbol that is #defined only if
PERL_CORE is defined won't be visible, and need not be on the exception
list.

This cuts almost 30% off the initial list.