Add grok_bin_hex()
grok_bin() and grok_hex() differ from grok_oct() in that they optionally
accept a leading 0x or 0b, based on a flag.  Then they all share common
code.  This commit splits out the check for the leading 0[bx] into a
separate function.

This speeds up the function common to all, since it doesn't have to
check for this possibility.  It slows these two down, as there is an
extra function layer.  The reason to do this is that future commits will
make the common function called in more places where speed matters.