pp_ctl.c: Comment on 2 exceptions in pp_goto()
Per discussion in GH #23969, neither of these two exceptions can be
reached from our test suite. The exceptions address the possibility of a
subroutine without a name, which is unlikely except possibly from XS
code.

The exceptions immediately preceding each of the two upon which we are
commenting have messages composed to include a subroutine's name.  These
exceptions can be reached from our test suite and are covered as of the
time of this commit.