Rename THING token when used for attributes
Prior to this, toke.c and perly.y used the token name `THING` for a
number of apparently totally unrelated reasons. While I haven't
attempted to fix most of them, I did find that the one used for
attribute lists is unrelated to the others, so I have renamed that to
something more sensible and meaningful.