perlapi: Use new 'apidoc_flag' lines
This replaces the clumsy method used previously by the new construct
introduced in the previous commit.

There are many other potential places to use this new feature; this gets
just the important ones (the ones having separate apidoc entries for the
flags) and a few others that I tried.  The others aren't important
currently, as the only current effect of adding them would be to add X<>
entries to perlapi; that could change.  Devel::PPPort can be enhanced to
take advantage of these as well.