grok_bin_oct_hex: Eke more precision when overflows
The algorithm for dealing with input that overflows a UV is to parse in
batches that don't overflow.  This means when we would otherwise
overflow the UV, we store the unoverflowed value in an NV, and start
over with the next batch, combining that with the first batch at the end
or when it too overflows.  This minimizes the number of floating point
operations.

But on platforms where a UV contains more bits than an NV significand,
that first batch when stored loses precision.  This is typical on a
64-bit machine using IEEE 754 double precision NVs, which have 53 bits
of precision.

This precision loss can be avoided by stopping accumulating the first
batch just before it would overflow the significand.  That effectively
is what this commit does.

Actually it is a bit more complicated.  Instead of stopping there, it
adds a checkpoint and goes on.  If the final result doesn't overflow the
UV, the checkpoint is simply ignored.  But if it does overflow, the code
returns to the checkpoint and creates a first batch from the data as of
there, and reparses those typically few bytes of input (11 bits is 2-3
hex digits) as the first bytes of the second batch.