grok_uint_by_base: Clean up loop some
As the comments say, this loop is "wonky".  But it doesn't have to be as
wonky as it is.  This commit reverses the meaning of some conditionals
so that when true they exit the loop.  And this allows the removal of a
'break' just at the end of the loop scope, making the loop much more
conventional.

In reversing the meaning, I split some off into more conditionals.  This
makes it easier to understand, and prepares for a future commit.