grok_uint_by_base: Optimize for more likely case
This changes the code to prefer the case where the input is a string
without underscores that contains 8 or fewer digits in the given base.
Doing so eliminates two conditionals from what I think is what the vast
majority of calls to this function look like.

This is accomplished in this commit by making the default: case for at
least 9 digits and adds a case: for 8.  There is no need for anything
then but to return for any string 8 or fewer digit characters long.

The default does try to handle leading zeros efficiently by having a
tight loop to strip them off.