grok_uint_by_base: Round discarded digits
When remaining digits are discarded instead of overflowing, this commit
causes the returned non-overflowed value to be rounded (towards even)
instead of truncated.

This will be useful when future commits cause this function to be called
to convert a string of numbers following a decimal point.  Instead of
truncating the final significant value, it will be rounded.