embed.fnc: Add string assertions for S_intuit-more
This makes sure the terminating character is a NUL.  This internal
function isn't documented as having that requirement, but that's always
the case in our test suite.  And functions it calls assume there is at
least one character in the input, so the assertion shouldn't be EPTRge,
and the test suite fails if it is EPTRgt.