Replace remaining hard tabs with 4 spaces
Earlier in this pull request we tidied up leading whitespace and
uncuddled else's, but only in the parts of the file that we were
planning to revise.  At this point, we can do some common-sense tidying
in the remaining parts of the file.

We leave a hard tab in one case where it's illustrating a 'make' command.  Get indents in better shape
(though leaving some else's uncuddled unchanged where that's more readable).