Refactor code into _internal_build_extension()
We take the code that was previously within the scope of the NO_MAKEFILE
label and place it in its own subroutine.

Two of the newly created internal functions, _use_Makefile_PL() and
_is_unix(), will now be called from within _internal_build_extension()
rather than directly from build_extension().

Two of the four possible places at which build_extension could
previously return -- thereby triggering progression to the next module
to be built in sequence -- now fall within the scope of
_internal_build_extension().  These have to be a "hard return," i.e.,
they have to return an undefined value.  At the point at which they are
called from within build_extension(), we now capture that (undefined)
return value and say 'return unless defined $rv;' in order to return
from build_extension() and proceed to the next module's processing.