Replace instance of 'goto LABEL' with subroutine call
This is the commit at which we finally solve the problem discussed in
https://github.com/Perl/perl5/issues/24015.  We replace the 'goto' call
with a call of _internal_build_extension() with the same arguments we
are already using in the standard cases.

Can be verified by configuring and running make test_prep, mangling
dist/Data-Dumper/Makefile to change the value of 'XS_VERSION' -- which
previously (at least in a post-d306795336 world) would have caused a
build failure -- then re-run 'make test_prep' which, during the rebuild,
will generate a new, corrected dist/Data-Dumper/Makefile and run to
completion.