Warn when `refalias` desynchronises a closure-captured variable
The `refalias` feature is currently experimental due to a number of
limitations of its implementation, most notably the effect it has on
closures such as the one illustrated in this unit test. Any change in
implementation to make this case function as intended would be a large
undertaking, so instead as a path towards being able to remove the
experimental status from this feature, we consider this behaviour
warning instead.