embed.pl: Handle case of multiple flags for an element
Some symbols are #defined in multiple places in the input; based
typically on different preprocessor conditionals.  We want to use the
definition which has the widest visibility.

This change showed that USE_STDIO had wrongly been undefined for the
past few commits in blead.  It is no longer actually ever defined by
perl.