Add Configure probe for size of intmax_t
Now that we require C99, we should be able to probe for this

Some internal operations can benefit from using the widest type possible
on the system, even if IVs and UVs don't use the full width.

This data is kept in <stdint.h>, which should always be available in
C99-compliant build systems.  Configure has a 'i_stdint', but it comes
much later in Configure than this does.  I tried moving it to earlier,
but then it relies on stuff that would have to get moved too, and I
think that stuff has to be later.

It gets complicated trying to deal with this value not existing, as
there is a separate unit for 'long long', which would have to be
considered.  Until and unless some platform comes along that this
doesn't work for, I just assume that this C99 construct will work as
expected.