ParseXS: add basic tests for INCLUDE(_COMMAND)
There is currently a single test each for these two keywords in
002-more.t, which try to actually compile and run the generated C code.

This commit:

- adds two similar tests to 001-basic.t, which use the same test include
files, but which only generate the C file and test the contents, not
compile and run it.

- Adds a test for nearly every INCLUDE/INCLUDE_COMMAND error message,
none of which were tested before. Only two errors remain untested.

1) "Note: the INCLUDE directive with a command is discouraged"

which warns about non-portable usage, and which thus is difficult to
test portably.

2) "Error: INCLUDE_COMMAND: cannot run command"

which is for when the command fails. Perl does complicated things,
including a possible warning on STDERR in the forked child if it fails
to exec, plus dying in the parent with the same error. So I don't see an
easy way to portably test this.

This commit adds a new file, t/XSloop.xsh, which has an INCLUDE line
which refers to itself.