ParseXS: regularise INCLUDE pipe error handling
If INCLUDE or INCLUDE_COMMAND runs a pipe, there is code in the handler
to check for the return status of the sub-process being non-zero (e.g.
non-zero exit, died with signal etc).

That handling was a bit odd. Instead of using $pxs->death(...), it was
manually doing 'print STDERR ...; exit(1)'. Also, the error message was
displaying $!, which is unconnected with the return code of the child.

This commit:

- makes it use the death() method. As a side-effect, this makes it
  display the correct line where the bad INCLUDE is located;
  previously it displayed the end of the current paragraph + 1.

- Includes the value of $? rather than $! in the error message. It just
  displays it as a 4-digit hex number rather than trying to decode the
  its various parts.

- Uses the standard error message prefix form "Error: KEYWORD: ...".

- Adds a test.