ParseXS: be stricter on REQUIRE arg
The argument to the REQUIRE: keyword is supposed to be a number:
NNN or MMM.NNN. The validating regex didn't do a closing $, so any junk
following the value was allowed through, e.g. 3.0.0 or 3.0XYZ. This
commit makes such variants an error.

This seems unlikely to cause any backcompat issues.

Also, the invalid version number immediately has a numeric comparison
applied to it, which since 5.42.0 has been generating a Perl warning
like:

    Argument "3.0.0" isn't numeric in numeric ge (>=) at ...

Also a quick grep of CPAN shows no obvious distros which would break
(the REQUIRE keyword isn't common).