ParseXS: tweak param-parsing code
In a couple of places there are checks on whether certain constructs
such as IN_OUT are disallowed due to a switch like -noinout.

Because these use blurt() to report the error, parsing will continue.
So this commit makes the parsing code still process the IN_OUT modifier
(or whatever) as normal, to reduce the risk of a cascade of weird
errors.