ParseXS: ban IN_OUT etc modifiers with length(s)
Disallow signatures like

    int
    foo(char *s, IN_OUT int length(s))

Most of the I/O modifiers make no sense with length(), and either
produced bad C code, or in recent Perls, generated an internal error.
There seems little benefit in getting a couple of cases to work for some
sort of obscure semantics. So this commit goes for a blanket ban.