ParseXS: length() tests: add type
Several tests which were expected to raise an error didn't include a type
for the length() pseudo-parameter. This was harmless, as the error that
was being tested for occurred before any check for a missing type.

But best for a test case to only include the one syntax error which is
actually being checked for.