ParseXS: ban length() on placeholder param
Make this an error:

    int
    foo(placeholder, int length(placeholder))

A parameter without a type is a placeholder: it's not a real variable,
so its length can't be retrieved. Prior to this commit, autocall would
pass an uninitialised variable to the C library function.

Also add a code comment explaining where checks like these are done